Rewrite spec to work similarly regardless of server API used
The compact index API does not work like this test describes, because
the compact index API only considers subsets of dependency names,
according to whats in the Gemfile, and only names dependent on those are
fetched. Since activesupport is not really a dependency remotely, it's
not considered. The full index of course does consider activesupport for
resolution since the full index is downloaded, so resolution differs.

I think this is fine and not worth fixing. Editing dependencies locally
is a bad practice, and it's fine to ignore the installed package upon
resolution. On top of this, once lockfile checksums are in place,
Bundler would probably complain in that case anyways.